Compile each *.cc file into *.o and then link
This will let us reuse compilation steps as we add more and more files to the
project; previously, we were compiling and linking in a single step, which would
rebuild the entire project (currently small) with any change to any file.

Also add sanitizer flags to both compilation and link steps.