Rename `cpp.yml` -> `main.yml`
Also rename the overall CI config from "C++" to "Build" and C++ can be one of
the jobs; this will let us amortize the high cost of booting a new VM or
starting a container, and run several jobs inside each OS image for efficiency.

Update the CI badge in the README accordingly.